Split `notification-routes` into container and component
