+++ ui/app/pages/routes/notification-routes/index.js
@@ -1,0 +1,1 @@
+export { default } from './notification-routes.container'
